pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "file-search-cli"
version = "0.1.0"
description = "A professional-grade Elasticsearch-backed file search CLI"
authors = [{name = "Elastic Propio"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
    "python-dotenv>=1.0",
    "elasticsearch>=8.13",
    "tenacity>=8.2",
    "pdfminer.six>=20231228",
    "rich>=13.7",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "python-multipart>=0.0.9"
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.7",
    "pytest>=8.2",
]

[tool.setuptools.packages.find]
where = ["src"]
